Added some documentation to the example classes
